Refactor to use parkingsystem
