+++ ParkingSystem/.idea/workspace.xml
@@ -4,32 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="4cb78f78-9c8b-44e9-9843-61206fcfa855" name="Changes" comment="Push the pom.xml">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/Address.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/Address.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/Car.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/Car.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/CarType.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/CarType.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/Customer.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/Customer.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/Money.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/Money.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/ParkingCharge.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/ParkingCharge.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/ParkingLot.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/ParkingLot.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/ParkingOffice.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/ParkingOffice.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsystem/ParkingSystem.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/edu/du/ict4305/parkingsytem/ParkingSystem.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/AddressTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/AddressTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/CarTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/CarTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/CustomerTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/CustomerTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/MoneyTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/MoneyTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/ParkingChargeTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/ParkingChargeTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/ParkingLotTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/ParkingLotTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsystem/ParkingOfficeTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/java/edu/du/ict4305/parkingsytem/ParkingOfficeTest.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/Address.class" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/Car.class" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/CarType.class" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/Customer.class" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/ParkingLot.class" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/target/classes/edu/du/ict4305/parkingsystem/ParkingSystem.class" beforeDir="false" />
-    </list>
+    <list default="true" id="4cb78f78-9c8b-44e9-9843-61206fcfa855" name="Changes" comment="Update the print statements to start printing the toString" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -199,7 +174,7 @@
       <workItem from="1666269548106" duration="32863000" />
       <workItem from="1667754671541" duration="9614000" />
       <workItem from="1668106355808" duration="201000" />
-      <workItem from="1668106591327" duration="1580000" />
+      <workItem from="1668106591327" duration="2067000" />
     </task>
     <task id="LOCAL-00001" summary="Test the Car class getter methods">
       <created>1665086935477</created>
@@ -299,7 +274,14 @@
       <option name="project" value="LOCAL" />
       <updated>1665101137927</updated>
     </task>
-    <option name="localTasksCounter" value="15" />
+    <task id="LOCAL-00015" summary="Update the print statements to start printing the toString">
+      <created>1668108752572</created>
+      <option name="number" value="00015" />
+      <option name="presentableId" value="LOCAL-00015" />
+      <option name="project" value="LOCAL" />
+      <updated>1668108752572</updated>
+    </task>
+    <option name="localTasksCounter" value="16" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -330,7 +312,8 @@
     <MESSAGE value="Remove the file from the project but add it to the right folder" />
     <MESSAGE value="Use this one instead" />
     <MESSAGE value="Push the pom.xml" />
-    <option name="LAST_COMMIT_MESSAGE" value="Push the pom.xml" />
+    <MESSAGE value="Update the print statements to start printing the toString" />
+    <option name="LAST_COMMIT_MESSAGE" value="Update the print statements to start printing the toString" />
     <option name="REFORMAT_BEFORE_PROJECT_COMMIT" value="true" />
   </component>
   <component name="XSLT-Support.FileAssociations.UIState">
