+++ ParkingSystem/pom.xml
@@ -18,6 +18,6 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>11</maven.compiler.target>
-        <exec.mainClass>edu.du.ict4305.parkingsytem.ParkingSystem</exec.mainClass>
+        <exec.mainClass>edu.du.ict4305.parkingsystem.ParkingSystem</exec.mainClass>
     </properties>
 </project>
