+++ ParkingSystem/src/main/java/edu/du/ict4305/parkingsystem/Address.java
@@ -3,7 +3,7 @@
  * @File: Address.java
  * @Instructor: Dr. Sherri Maciosek
  */
-package edu.du.ict4305.parkingsytem;
+package edu.du.ict4305.parkingsystem;
 
 /**
  * The type Address.
