+++ ParkingSystem/src/main/java/edu/du/ict4305/parkingsystem/Car.java
@@ -3,7 +3,7 @@
  * @File: Car.java
  * @Instructor: Dr. Sherri Maciosek
  */
-package edu.du.ict4305.parkingsytem;
+package edu.du.ict4305.parkingsystem;
 
 import java.time.LocalDate;
 import java.util.Objects;
