+++ ParkingSystem/src/main/java/edu/du/ict4305/parkingsystem/Money.java
@@ -3,7 +3,7 @@
  * @Project: ParkingSystem
  * @Instructor: Dr. Sherri Maciosek
  */
-package edu.du.ict4305.parkingsytem;
+package edu.du.ict4305.parkingsystem;
 
 /**
  * The type Money.
