+++ ParkingSystem/src/main/java/edu/du/ict4305/parkingsystem/ParkingSystem.java
@@ -3,7 +3,7 @@
  * @File: ParkingSystem.java
  * @Instructor: Dr. Sherri Maciosek
  */
-package edu.du.ict4305.parkingsytem;
+package edu.du.ict4305.parkingsystem;
 
 import java.time.Instant;
 import java.time.LocalDate;
