+++ ParkingSystem/src/test/java/edu/du/ict4305/parkingsystem/ParkingChargeTest.java
@@ -3,7 +3,7 @@
  * @Project: ParkingSystem
  * @Instructor: Dr. Sherri Maciosek
  */
-package edu.du.ict4305.parkingsytem;
+package edu.du.ict4305.parkingsystem;
 import org.junit.jupiter.api.Test;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
